# api/openapi.yaml
openapi: 3.1.0
info:
  title: meds-api
  version: "1.0.0"
  summary: API REST en lecture seule sur la Base de données publique des médicaments (BDPM).
  description: |
    API REST exposant la **Base de données publique des médicaments** (BDPM), publiée par l'ANSM
    avec le concours de la HAS et du ministère de la Santé.

    ## Avertissement

    Les données proviennent de la Base de données publique des médicaments (ANSM), diffusée sous
    Licence Ouverte (Etalab). Elles sont fournies **à titre informatif** et **ne se substituent ni
    au Résumé des Caractéristiques du Produit, ni à l'avis d'un professionnel de santé**. Cette API
    ne constitue pas un dispositif médical et n'apporte aucune aide à la décision clinique.
    La date de mise à jour du jeu de données est exposée sur `/v1/dataset`.

    ## Conventions

    - Toutes les routes `/v1/*` exigent une clé API (`Authorization: Bearer <clé>`).
    - Les montants sont des **entiers en centimes**, suffixés `_cents`.
    - L'absence de valeur est toujours `null`, jamais `""` ni `0`. La distinction compte :
      7 276 des 20 903 présentations n'ont aucun prix, et les confondre avec un prix nul
      fausserait toute statistique.
    - Les valeurs métier sont **conservées telles quelles** (`Autorisation active`,
      `Commercialisée`), sans traduction ni normalisation.
    - Les erreurs suivent la **RFC 9457** (`application/problem+json`).
  contact:
    name: Support meds-api
...
  license:
...
servers:
  - url: /
...
security:
  - bearerAuth: []
...
paths:
...
  /v1/medicaments/{cis}:
    get:
      tags: [Médicaments]
      summary: Fiche d'un médicament
      description: |
        Sans `include`, seule la spécialité est renvoyée. Ce choix est délibéré : la fiche complète
        d'un médicament à 37 présentations et 40 avis pèse plusieurs dizaines de kilooctets,
        inutiles à un appelant qui ne veut qu'un libellé.
      operationId: getMedicament
      parameters:
        - $ref: "#/components/parameters/CIS"
        - name: include
          in: query
          description: Sections à inclure, séparées par des virgules, ou `all`.
          schema:
            type: string
            examples: ["presentations,composition", "all"]
      responses:
        "200":
          description: Fiche du médicament.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data: { $ref: "#/components/schemas/Medicament" }
                  meta: { $ref: "#/components/schemas/MetaSimple" }
        "304": { $ref: "#/components/responses/NotModified" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/NotReady" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Clé API d'au moins 32 caractères.
    adminAuth:
      type: http
      scheme: bearer
      description: Clé d'administration, distincte des clés API.
  parameters:
    CIS:
      name: cis
      in: path
      required: true
      description: Code identifiant de spécialité, huit chiffres.
      schema: { type: string, pattern: "^[0-9]{8}$" }
      example: "60234100"
...
  headers:
    ETag:
      description: Empreinte de la réponse, à renvoyer dans `If-None-Match`.
      schema: { type: string }
...
  responses:
    NotModified:
      description: La ressource n'a pas changé depuis l'`ETag` fourni.
    BadRequest:
      description: Paramètre absent, mal formé ou hors bornes.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unauthorized:
      description: Clé API absente ou invalide.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: Ressource inexistante.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    NotReady:
      description: Aucun jeu de données chargé.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      description: Erreur au format RFC 9457.
      required: [type, title, status, request_id]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        request_id: { type: string }
      example:
        type: https://meds-api.example/errors/not_found
        title: Ressource introuvable
        status: 404
        detail: Aucune spécialité ne porte le code CIS 99999999.
        instance: /v1/medicaments/99999999
        request_id: 01J8Z9K2M3N4P5Q6R7S8T9V0W1
    MetaSimple:
      type: object
      properties:
        dataset_version: { type: string }
...
    Medicament:
      type: object
      required: [cis, denomination]
      properties:
        cis: { type: string }
        denomination: { type: string }
        forme_pharmaceutique: { type: string }
        voies_administration: { type: array, items: { type: string } }
        statut_amm: { type: string }
        procedure_amm: { type: string }
        etat_commercialisation: { type: string }
        date_amm: { type: [string, "null"], format: date }
        statut_bdm: { type: [string, "null"] }
        numero_autorisation_europeenne: { type: [string, "null"] }
        titulaires: { type: array, items: { type: string } }
        surveillance_renforcee: { type: boolean }
        mitm:
          oneOf:
            - { $ref: "#/components/schemas/MITM" }
            - { type: "null" }
        presentations: { type: array, items: { $ref: "#/components/schemas/Presentation" } }
        composition: { type: array, items: { $ref: "#/components/schemas/Composant" } }
        generiques:
          oneOf:
            - { $ref: "#/components/schemas/GroupeGenerique" }
            - { type: "null" }
        avis: { $ref: "#/components/schemas/Avis" }
        conditions: { type: array, items: { type: string } }
        ruptures: { type: array, items: { $ref: "#/components/schemas/Rupture" } }
    Presentation:
      type: object
      required: [cip13, cip7, libelle]
      properties:
        cip13: { type: string }
        cip7: { type: string }
        libelle: { type: string }
        statut_administratif: { type: string }
        etat_commercialisation: { type: string }
        date_declaration: { type: [string, "null"], format: date }
        agrement_collectivites: { type: [boolean, "null"] }
        taux_remboursement: { type: array, items: { type: integer } }
        prix_medicament_cents:
          type: [integer, "null"]
          description: Prix en centimes. `null` si absent — jamais confondu avec zéro.
        prix_public_cents: { type: [integer, "null"] }
        honoraires_dispensation_cents: { type: [integer, "null"] }
        indications_remboursement: { type: [string, "null"] }
...
    Composant:
      type: object
      properties:
        element_pharmaceutique: { type: string }
        code_substance: { type: string }
        denomination_substance: { type: string }
        dosage:
          type: string
          description: |
            Texte brut, jamais interprété : la source mêle notations métriques et homéopathiques,
            et normaliser serait interpréter une donnée de santé.
        reference_dosage: { type: string }
        nature: { type: string, description: "`SA` (substance active) ou `FT` (fraction thérapeutique)." }
        numero_liaison: { type: integer }
    Avis:
      type: object
      properties:
        smr: { type: array, items: { $ref: "#/components/schemas/AvisItem" } }
        asmr: { type: array, items: { $ref: "#/components/schemas/AvisItem" } }
    AvisItem:
      type: object
      properties:
        code_dossier_has: { type: string }
        motif_evaluation: { type: string }
        date_avis: { type: [string, "null"], format: date }
        valeur: { type: string }
        niveau:
          type: [integer, "null"]
          description: "Niveau ASMR de 1 à 5. `null` si la valeur n'est pas un chiffre romain pur — la clé est toujours présente."
        libelle: { type: string }
        lien_avis_ct: { type: [string, "null"], format: uri }
    GroupeGenerique:
      type: object
      properties:
        id: { type: string }
        libelle: { type: string }
        membres:
          type: array
          items:
            type: object
            properties:
              cis: { type: string }
              denomination: { type: string }
              type:
                type: integer
                description: "Énumération discontinue : 0, 1, 2, 4 — la valeur 3 n'existe pas."
                enum: [0, 1, 2, 4]
              type_libelle: { type: string }
              ordre: { type: integer }
              etat_commercialisation: { type: string }
              _link: { type: string }
    Rupture:
      type: object
      properties:
        cis: { type: string }
        cip13:
          type: [string, "null"]
          description: "`null` signifie que toute la spécialité est concernée."
        code_statut: { type: integer }
        statut: { type: string, enum: [rupture, tension, arret, remise_disposition] }
        libelle: { type: string, description: Libellé dérivé du code, jamais repris de la source. }
        libelle_source: { type: string, description: Libellé brut de la source, conservé pour l'audit. }
        date_debut: { type: [string, "null"], format: date }
        date_debut_approximative: { type: boolean }
        date_mise_a_jour: { type: [string, "null"], format: date }
        date_remise_disposition: { type: [string, "null"], format: date }
        lien_ansm: { type: string, format: uri }
    MITM:
      type: object
      properties:
        code_atc: { type: string }
        denomination: { type: string }
        lien_bdpm: { type: string, format: uri }
...